Refactor extention header into library
