+++ python/ecole/src/ecole/__init__.py
