+++ python/ecole/src/ecole/core/caster.hpp
