+++ python/ecole/src/ecole/core/core.cpp
