+++ python/ecole/src/ecole/core/core.hpp
