+++ python/ecole/src/ecole/core/data.cpp
