+++ python/ecole/src/ecole/core/dynamics.cpp
