+++ python/ecole/src/ecole/core/information.cpp
