+++ python/ecole/src/ecole/core/instance.cpp
