+++ python/ecole/src/ecole/core/reward.cpp
