+++ python/ecole/src/ecole/core/scip.cpp
