+++ python/ecole/src/ecole/data.py
