+++ python/ecole/src/ecole/doctor.py
