+++ python/ecole/src/ecole/environment.py
