+++ python/ecole/src/ecole/information.py
