+++ python/ecole/src/ecole/instance.py
