+++ python/ecole/src/ecole/observation.py
