+++ python/ecole/src/ecole/reward.py
