+++ python/ecole/src/ecole/scip.py
