+++ python/ecole/src/ecole/typing.py
