+++ python/ecole/src/ecole/version.py
