+++ python/ecole/tests/_test_instance_old.py
