+++ python/ecole/tests/test_data.py
