+++ python/ecole/tests/test_environment.py
