+++ python/ecole/tests/test_information.py
