+++ python/ecole/tests/test_instance.py
