+++ python/ecole/tests/test_observation.py
