+++ python/ecole/tests/test_reward.py
