+++ python/ecole/tests/test_scip.py
