+++ python/ecole/tests/test_version.py
