+++ python/extension-helper/include/ecole/python/auto-class.hpp
